requirements.txt
torch
torchvision
os
Pillow
streamlit